Move sync events to V3 API packages
