+++ cardinal-components-chunk/src/main/java/dev/onyxstudios/cca/api/v3/chunk/ChunkSyncCallback.java
@@ -20,10 +20,26 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-@MethodsReturnNonnullByDefault
-@ParametersAreNonnullByDefault
-package nerdhub.cardinal.components.api.event;
+package dev.onyxstudios.cca.api.v3.chunk;
 
-import nerdhub.cardinal.components.api.util.MethodsReturnNonnullByDefault;
+import net.fabricmc.fabric.api.event.Event;
+import net.fabricmc.fabric.api.event.EventFactory;
+import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.world.chunk.WorldChunk;
 
-import javax.annotation.ParametersAreNonnullByDefault;
+/**
+ * The callback interface for receiving chunk synchronization events.
+ */
+@FunctionalInterface
+public interface ChunkSyncCallback {
+    Event<ChunkSyncCallback> EVENT = EventFactory.createArrayBacked(ChunkSyncCallback.class, (p, chunk) -> {}, listeners -> (player, chunk) -> {
+        for (ChunkSyncCallback callback : listeners) {
+            callback.onChunkSync(player, chunk);
+        }
+    });
+
+    /**
+     * Called when a chunk's data is sent to a player
+     */
+    void onChunkSync(ServerPlayerEntity player, WorldChunk chunk);
+}
