+++ cardinal-components-chunk/src/main/java/dev/onyxstudios/cca/mixin/chunk/common/MixinThreadedAnvilChunkStorage.java
@@ -23,7 +23,7 @@
 package dev.onyxstudios.cca.mixin.chunk.common;
 
 import com.mojang.datafixers.DataFixer;
-import nerdhub.cardinal.components.api.event.ChunkSyncCallback;
+import dev.onyxstudios.cca.api.v3.chunk.ChunkSyncCallback;
 import net.minecraft.network.Packet;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ThreadedAnvilChunkStorage;
