+++ cardinal-components-entity/src/main/resources/fabric.mod.json
@@ -8,10 +8,10 @@
     "side": "universal",
     "entrypoints": {
         "main": [
-          "nerdhub.cardinal.components.CardinalComponentsEntity::init"
+          "dev.onyxstudios.cca.internal.entity.CardinalComponentsEntity::init"
         ],
         "client": [
-            "nerdhub.cardinal.components.CardinalComponentsEntity::initClient"
+          "dev.onyxstudios.cca.internal.entity.CardinalComponentsEntity::initClient"
         ]
     },
     "custom": {
